rename handlers package to handler
