+++ internal/handler/regenerate.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"encoding/json"
